Move SqlQuery content into own file
