+++ packages/database-editor/src/components/editor/master/MasterControl.tsx
@@ -49,7 +49,7 @@ export const MasterControl = ({ table, deleteDatabaseConnection, selectedDatabas
       </TooltipProvider>
       <Separator decorative orientation='vertical' style={{ height: '20px', margin: 0 }} />
       <SqlQueryTester selectedDatabase={selectedDatabase}>
-        <Button disabled={table.getSelectedRowModel().flatRows.length === 0} icon={IvyIcons.Configuration} />
+        <Button disabled={table.getSelectedRowModel().flatRows.length === 0} icon={IvyIcons.Sql} />
       </SqlQueryTester>
       <Separator decorative orientation='vertical' style={{ height: '20px', margin: 0 }} />
       <ImportWizard context={{ app: context.app, file: context.file, projects: projects }}>
